Create MangaDatabaseInitializer.py
This is a utility for MangaMine to download all the manga names and links from batoto, mangahere, and mangapanda (Mangastream is not supported here, sorry). Simply run this in the same directory as MangaMine.py and then when it is completed do not touch the folder and Database.txt file that it creates. MangaMine will detect the folder and .txt file.

This program is meant to be run weekly, biweekly or once a month to keep an updated list of all manga for each website available to use instead of haing to find and copy and paste links yourself.

DO NOT ALTER THE DATABASE TEXT FILE. DOING SO WILL BREAK MANGAMINE. ALTER ONLY AT YOUR OWN RISK.