Updated The MangaStream Portion
I forgot to add in a portion that if the number of objects in the current working directory equal the number of pages in the chapter, it will skip trying to download that chapter and go to the next one. This will save time.

These few lines of code (similar in all 4 site functions) assume that the folders wont be altered in anyway. If they are the program will just go on and try to down the images, and if there is a file by that name and extension it will skip that then go to the next page.